feat: Track post collaborators with an array
Adds a `collaboratorIds` array to the `Post` type and initializes it in the `InitiativeHub` and `CommunityPage` components. This allows tracking which users are collaborating on a post. The `PostCard` component is updated to display collaboration status and enable toggling participation by adding/removing the user's ID to/from the `collaboratorIds` array.